+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased Changes
 * Added a lint to prevent scripts in maps
 * Added Map Items, Map NPCSpawns, Map Clipping sections
+* Added a lint to tagged ensure items are descendants of the items folder
 * Changed the color of Clip_Player to differ from Clip_Bounds
 * Changed the PanelGroup dropdown visual
 * Fixed entrance linting throwing an error if a node wasn't found